+++ lib/pipas_persister/service.rb
@@ -10,6 +10,7 @@ require_relative 'service/patients'
 require_relative 'service/appointments'
 require_relative 'service/deliveries'
 require_relative 'service/treatments'
+require_relative 'service/treatment_plans'
 require_relative 'service/service_info'
 require_relative 'service/facade'
 require_relative 'service/protocol'
@@ -52,6 +53,10 @@ module PipasPersister
 
       map '/patients' do
         run Service::Patients
+      end
+
+      map '/treatment-plans' do
+        run Service::TreatmentPlans
       end
 
       map '/treatments' do